Refactor: carve out testing utils
